+++ src/app/00043-CuentasPorCobrar/Data/Tables/TC_DependenciaUNFV.cs
@@ -76,7 +76,7 @@ namespace Data.Tables
             {
                 using (var _dbConnection = new SqlConnection(Database.ConnectionString))
                 {
-                    parameters.Add(name: "I_DependenciaID", dbType: DbType.Byte, value: this.I_DependenciaID);
+                    parameters.Add(name: "I_DependenciaID", dbType: DbType.Int32, value: this.I_DependenciaID);
                     parameters.Add(name: "B_Habilitado", dbType: DbType.Boolean, value: this.B_Habilitado);
                     parameters.Add(name: "D_FecMod", dbType: DbType.DateTime, value: this.D_FecMod);
                     parameters.Add(name: "CurrentUserId", dbType: DbType.Int32, value: currentUserId);
