+++ src/app/00043-CuentasPorCobrar/Data/Tables/TC_EntidadFinanciera.cs
@@ -75,7 +75,7 @@ namespace Data.Tables
             {
                 using (var _dbConnection = new SqlConnection(Database.ConnectionString))
                 {
-                    parameters.Add(name: "I_EntidadFinanID", dbType: DbType.Byte, value: this.I_EntidadFinanID);
+                    parameters.Add(name: "I_EntidadFinanID", dbType: DbType.Int32, value: this.I_EntidadFinanID);
                     parameters.Add(name: "B_Habilitado", dbType: DbType.Boolean, value: this.B_Habilitado);
                     parameters.Add(name: "D_FecMod", dbType: DbType.DateTime, value: this.D_FecMod);
                     parameters.Add(name: "CurrentUserId", dbType: DbType.Int32, value: currentUserId);
@@ -106,7 +106,7 @@ namespace Data.Tables
             {
                 using (var _dbConnection = new SqlConnection(Database.ConnectionString))
                 {
-                    parameters.Add(name: "I_EntidadFinanID", dbType: DbType.Byte, value: this.I_EntidadFinanID);
+                    parameters.Add(name: "I_EntidadFinanID", dbType: DbType.Int32, value: this.I_EntidadFinanID);
                     parameters.Add(name: "T_EntidadDesc", dbType: DbType.String, size: 250, value: this.T_EntidadDesc);
                     parameters.Add(name: "B_Habilitado", dbType: DbType.Boolean, value: this.B_Habilitado);
                     parameters.Add(name: "B_Archivos", dbType: DbType.Boolean, value: enableFiles);
@@ -140,7 +140,7 @@ namespace Data.Tables
             {
                 using (var _dbConnection = new SqlConnection(Database.ConnectionString))
                 {
-                    parameters.Add(name: "I_EntidadFinanID", dbType: DbType.Byte, value: this.I_EntidadFinanID);
+                    parameters.Add(name: "I_EntidadFinanID", dbType: DbType.Int32, value: this.I_EntidadFinanID);
                     parameters.Add(name: "T_EntidadDesc", dbType: DbType.String, size: 250, value: this.T_EntidadDesc);
                     parameters.Add(name: "B_Habilitado", dbType: DbType.Boolean, value: this.B_Habilitado);
                     parameters.Add(name: "D_FecMod", dbType: DbType.DateTime, value: this.D_FecMod);
