+++ src/app/00043-CuentasPorCobrar/Data/Tables/TS_CorreoAplicacion.cs
@@ -77,7 +77,7 @@ namespace Data.Tables
             {
                 using (var _dbConnection = new SqlConnection(Database.ConnectionString))
                 {
-                    parameters.Add(name: "I_CorreoID", dbType: DbType.Byte, value: this.I_CorreoID);
+                    parameters.Add(name: "I_CorreoID", dbType: DbType.Int32, value: this.I_CorreoID);
                     parameters.Add(name: "T_DireccionCorreo", dbType: DbType.String, size: 50, value: this.T_DireccionCorreo);
                     parameters.Add(name: "T_PasswordCorreo", dbType: DbType.String, size: 500, value: this.T_PasswordCorreo);
                     parameters.Add(name: "T_Seguridad", dbType: DbType.String, size: 50, value: this.T_Seguridad);
@@ -113,7 +113,7 @@ namespace Data.Tables
             {
                 using (var _dbConnection = new SqlConnection(Database.ConnectionString))
                 {
-                    parameters.Add(name: "I_CorreoID", dbType: DbType.Byte, value: this.I_CorreoID);
+                    parameters.Add(name: "I_CorreoID", dbType: DbType.Int32, value: this.I_CorreoID);
                     parameters.Add(name: "B_Habilitado", dbType: DbType.Boolean, value: this.B_Habilitado);
                     parameters.Add(name: "D_FecUpdate", dbType: DbType.DateTime, value: this.D_FecUpdate);
                     parameters.Add(name: "CurrentUserId", dbType: DbType.Int32, value: currentUserId);
@@ -144,7 +144,7 @@ namespace Data.Tables
             {
                 using (var _dbConnection = new SqlConnection(Database.ConnectionString))
                 {
-                    parameters.Add(name: "I_CorreoID", dbType: DbType.Byte, value: this.I_CorreoID);
+                    parameters.Add(name: "I_CorreoID", dbType: DbType.Int32, value: this.I_CorreoID);
                     parameters.Add(name: "T_DireccionCorreo", dbType: DbType.String, size: 50, value: this.T_DireccionCorreo);
                     parameters.Add(name: "T_PasswordCorreo", dbType: DbType.String, size: 500, value: this.T_PasswordCorreo);
                     parameters.Add(name: "T_Seguridad", dbType: DbType.String, size: 50, value: this.T_Seguridad);
